+++ field/convertMap.py
@@ -56,7 +56,7 @@ def run(
     createRootMap(inFileName, rootFileName, cmScale, storeCoords)
 
 
-def createRootMap(inFileName, rootFileName, cmScale, storeCoords) -> None:
+def createRootMap(inFileName: str, rootFileName: str, cmScale: float, storeCoords: bool) -> None:
     print(f"Create map {rootFileName} from {inFileName} using cmScale = {cmScale}")
     if storeCoords is True:
         print(f"We will also store the x,y,z field coordinates in {rootFileName}")
@@ -165,7 +165,7 @@ def createRootMap(inFileName, rootFileName, cmScale, storeCoords) -> None:
     theFile.Close()
 
 
-def findRanges(inFileName, cmScale):
+def findRanges(inFileName: str, cmScale: float):
     # First read the data file to find the binning and coordinate ranges.
     # Store the unique (ordered) x, y and z values so we can then find the
     # bin widths, min/max limits and central offset
@@ -201,6 +201,12 @@ def findRanges(inFileName, cmScale):
     xMin = 0.0
     xMax = 0.0
     dx = 0.0
+    yMin = 0.0
+    yMax = 0.0
+    dy = 0.0
+    zMin = 0.0
+    zMax = 0.0
+    dz = 0.0
 
     if Nx > 0:
         xMin = xArray[0]